Join AccessInterceptorScopeLocalizerAllowUseVariadicNotationTest with AccessInterceptorScopeLocalizerFunctionalTest, and fix return on Util class.
